Modify folder name
