--- weevent-jemeter/src/main/java/com/webank/weevent/jemeter/JMeterApplication.java
@@ -1,13 +1,0 @@
-package com.webank.weevent.jemeter;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class JMeterApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(JMeterApplication.class, args);
-    }
-
-}
