+++ weevent-jemeter/src/main/java/com/webank/weevent/jmeter/producer/WeEventProducer.java
@@ -1,39 +1,48 @@
-package com.webank.weevent.jemeter.consumer;
+package com.webank.weevent.jmeter.producer;
 
 
+import com.webank.weevent.sdk.BrokerException;
 import com.webank.weevent.sdk.IWeEventClient;
+import com.webank.weevent.sdk.SendResult;
 import com.webank.weevent.sdk.WeEvent;
 
-import lombok.extern.slf4j.Slf4j;
 import org.apache.jmeter.config.Arguments;
 import org.apache.jmeter.protocol.java.sampler.AbstractJavaSamplerClient;
 import org.apache.jmeter.protocol.java.sampler.JavaSamplerContext;
 import org.apache.jmeter.samplers.SampleResult;
-import org.springframework.beans.factory.annotation.Autowired;
 
 /**
- * WeEvent consumer performance.
+ * WeEvent publish performance.
  *
- * @author matthewliu@webank.com
+ * @author puremilkfan
  * @since 2018/11/27
  */
-@Slf4j
-public class WeEventConsumer extends AbstractJavaSamplerClient {
+public class WeEventProducer extends AbstractJavaSamplerClient {
     private String topic;
-    @Autowired
-    private static IWeEventClient weEventClient;
-
+    private String buffer;
+    private IWeEventClient weEventClient;
 
     // 每个压测线程启动时跑一次
     @Override
     public void setupTest(JavaSamplerContext context) {
         super.setupTest(context);
+
         try {
             this.topic = context.getParameter("topic");
+            int size = context.getIntParameter("size_byte");
+            getNewLogger().info("params topic: {}, size: {}kb", this.topic, size);
+
+            StringBuilder sb = new StringBuilder();
+            for (int i = 0; i < size; i++) {
+                // 1kb
+                sb.append("a");
+            }
+            this.buffer = sb.toString();
+
             boolean result = this.weEventClient.open(this.topic);
             getNewLogger().info("open topic result: {}", result);
-        } catch (Exception e) {
-            getNewLogger().error("open Exception", e);
+        } catch (BrokerException e) {
+            getNewLogger().error("open ClientException", e);
         }
     }
 
@@ -48,6 +57,7 @@ public class WeEventConsumer extends AbstractJavaSamplerClient {
     public Arguments getDefaultParameters() {
         Arguments arguments = new Arguments();
         arguments.addArgument("topic", "com.webank.weevent.jmeter");
+        arguments.addArgument("size_byte","1");
         return arguments;
     }
 
@@ -55,22 +65,13 @@ public class WeEventConsumer extends AbstractJavaSamplerClient {
     @Override
     public SampleResult runTest(JavaSamplerContext context) {
         SampleResult result = new SampleResult();
-        result.setSampleLabel("weEventClient");
+        result.setSampleLabel("publish");
         try {
             result.sampleStart();
-            this.weEventClient.subscribe(this.topic, WeEvent.OFFSET_LAST, new IWeEventClient.EventListener() {
-                @Override
-                public void onEvent(WeEvent event) {
-
-                }
-
-                @Override
-                public void onException(Throwable e) {
-
-                }
-            });
+            SendResult sendResult = this.weEventClient.publish(new WeEvent(this.topic, this.buffer.getBytes()));
             result.sampleEnd();
-            result.setSuccessful(true);
+            result.setSuccessful(sendResult.getStatus() == SendResult.SendResultStatus.SUCCESS && sendResult.getEventId().length() > 0);
+            result.setResponseMessage(sendResult.getEventId());
         } catch (Exception e) {
             getNewLogger().error("publish Exception", e);
             result.sampleEnd();
@@ -78,27 +79,5 @@ public class WeEventConsumer extends AbstractJavaSamplerClient {
             result.setResponseMessage(e.getMessage());
         }
         return result;
-    }
-
-    public static void main(String[] args) {
-        String topic = "com.webank.irps.002";
-        try {
-            weEventClient.subscribe(topic, WeEvent.OFFSET_LAST, new IWeEventClient.EventListener() {
-                @Override
-                public void onEvent(WeEvent event) {
-                    log.info(String.format("onEvent %s", event));
-                }
-
-                @Override
-                public void onException(Throwable e) {
-                    log.info(String.format("onException %s", e.getMessage()));
-                    e.printStackTrace();
-                }
-            });
-
-            Thread.sleep(10000);
-        } catch (Exception e) {
-            log.error("subscribe fail,error:{}",e.getMessage());
-        }
     }
 }
